backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
Pillow>=9.0.0
torch>=2.0.0
torchvision>=0.15.0
timm==1.0.24
